# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrewGo - Beverage ordering platform for Bangalore cafes
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Cafe {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  area: string;
  address: string;
  rating: number;
  reviewCount?: number;
  /** Estimated delivery time in minutes */
  deliveryTime: number;
  deliveryFee: number;
  minOrder: number;
  imageUrl: string;
  categories: string[];
  isOpen: boolean;
  totalOrders: number;
  isFeatured?: boolean;
  /**
     * e.g. "20% off on first order"
     * @nullable
     */
  discount?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
}

export interface CafeInput {
  name: string;
  description?: string;
  area: string;
  address: string;
  deliveryTime: number;
  deliveryFee: number;
  minOrder: number;
  imageUrl: string;
  categories: string[];
  isFeatured?: boolean;
  discount?: string;
  latitude?: number;
  longitude?: number;
}

export type MenuItemCategory = typeof MenuItemCategory[keyof typeof MenuItemCategory];


export const MenuItemCategory = {
  coffee: 'coffee',
  tea: 'tea',
  smoothie: 'smoothie',
  mojito: 'mojito',
  beverage: 'beverage',
} as const;

export interface MenuItem {
  id: number;
  cafeId: number;
  cafeName: string;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: MenuItemCategory;
  imageUrl: string;
  isAvailable: boolean;
  isVeg: boolean;
  isBestseller?: boolean;
  /**
     * JSON string of customization options
     * @nullable
     */
  customizations?: string | null;
}

export interface Review {
  id: number;
  cafeId: number;
  rating: number;
  comment: string;
  reviewerName: string;
  createdAt: string;
}

export interface ReviewInput {
  rating: number;
  comment: string;
  reviewerName: string;
}

export interface OrderItem {
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
  imageUrl: string;
  /** @nullable */
  customization?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  placed: 'placed',
  confirmed: 'confirmed',
  preparing: 'preparing',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  cafeId: number;
  cafeName: string;
  items: OrderItem[];
  status: OrderStatus;
  subtotal: number;
  deliveryFee: number;
  total: number;
  deliveryAddress: string;
  /** Estimated delivery time remaining in minutes */
  estimatedTime: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface OrderItemInput {
  menuItemId: number;
  quantity: number;
  customization?: string;
}

export interface OrderInput {
  cafeId: number;
  items: OrderItemInput[];
  deliveryAddress: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  placed: 'placed',
  confirmed: 'confirmed',
  preparing: 'preparing',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface Banner {
  id: number;
  title: string;
  subtitle: string;
  imageUrl: string;
  color: string;
  /** @nullable */
  ctaText?: string | null;
}

export interface Offer {
  id: number;
  title: string;
  code: string;
  discount: string;
  description: string;
}

export interface FeaturedContent {
  banners: Banner[];
  featuredCafes: Cafe[];
  offers: Offer[];
}

export interface Area {
  name: string;
  cafeCount: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Category {
  name: string;
  slug: string;
  itemCount: number;
  imageUrl: string;
}

export interface TrendingContent {
  topCafes: Cafe[];
  popularItems: MenuItem[];
}

export type ListCafesParams = {
area?: string;
category?: string;
search?: string;
sortBy?: ListCafesSortBy;
};

export type ListCafesSortBy = typeof ListCafesSortBy[keyof typeof ListCafesSortBy];


export const ListCafesSortBy = {
  rating: 'rating',
  deliveryTime: 'deliveryTime',
  popularity: 'popularity',
} as const;

export type SearchMenuItemsParams = {
q?: string;
category?: string;
};