# lib/api-zod/src/generated/types/cafe.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrewGo - Beverage ordering platform for Bangalore cafes
 * OpenAPI spec version: 0.1.0
 */

export interface Cafe {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  area: string;
  address: string;
  rating: number;
  reviewCount?: number;
  /** Estimated delivery time in minutes */
  deliveryTime: number;
  deliveryFee: number;
  minOrder: number;
  imageUrl: string;
  categories: string[];
  isOpen: boolean;
  totalOrders: number;
  isFeatured?: boolean;
  /**
     * e.g. "20% off on first order"
     * @nullable
     */
  discount?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
}