# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrewGo - Beverage ordering platform for Bangalore cafes
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemCategory } from './menuItemCategory';

export interface MenuItem {
  id: number;
  cafeId: number;
  cafeName: string;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: MenuItemCategory;
  imageUrl: string;
  isAvailable: boolean;
  isVeg: boolean;
  isBestseller?: boolean;
  /**
     * JSON string of customization options
     * @nullable
     */
  customizations?: string | null;
}