# lib/api-zod/src/generated/types/offer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrewGo - Beverage ordering platform for Bangalore cafes
 * OpenAPI spec version: 0.1.0
 */

export interface Offer {
  id: number;
  title: string;
  code: string;
  discount: string;
  description: string;
}