# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrewGo - Beverage ordering platform for Bangalore cafes
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  cafeId: number;
  rating: number;
  comment: string;
  reviewerName: string;
  createdAt: string;
}