# lib/api-zod/src/generated/types/trendingContent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrewGo - Beverage ordering platform for Bangalore cafes
 * OpenAPI spec version: 0.1.0
 */
import type { Cafe } from './cafe';
import type { MenuItem } from './menuItem';

export interface TrendingContent {
  topCafes: Cafe[];
  popularItems: MenuItem[];
}